LayoutTests/fast/css/pseudo-valid-006.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../js/resources/js-test-style.css">
<script src="../js/resources/js-test-pre.js"></script>
<style>
 :valid { background: red; }
 :invalid { background: red; }
 button { background: lime; }
 input { background: lime; }
</style>
</head>
<body>
<p id="description"></p>
<form method="get">
<button name="victim">Lorem ipsum</button>
<input name="victim" type="reset" value="Lorem ipsum"/>
<input name="victim" type="hidden" value="Lorem ipsum"/>
</form>
<div id="console"></div>
<script>
description("This test performs a check for the :valid/:invalid CSS selector on elements where it shouldn't be applied, part two.");

v = document.getElementsByName("victim");
for (i = 0; i < v.length; i++)
    shouldBe("document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color')", "'rgb(0, 255, 0)'");

var successfullyParsed = true;
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
