LayoutTests/fast/css/pseudo-valid-006-expected.txt

This test performs a check for the :valid/:invalid CSS selector on elements where it shouldn't be applied, part two.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Lorem ipsum 
PASS document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color') is 'rgb(0, 255, 0)'
PASS document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color') is 'rgb(0, 255, 0)'
PASS document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color') is 'rgb(0, 255, 0)'
PASS successfullyParsed is true

TEST COMPLETE
